Always set non-zero http timeouts

+++ jwt-client/jwt-client-core/src/main/java/org/entur/jwt/client/properties/JwtClientCache.java
@@ -5,10 +5,10 @@ public class JwtClientCache {
     private boolean enabled = true;
 
     /** Minimum number of seconds left before the returned (cached) token expires */
-    private int minimumTimeToLive = 15; // in seconds
+    private long minimumTimeToLive = 15; // in seconds
 
     /** Number of seconds to wait for a token, if it must be refreshed */
-    private int refreshTimeout = 15; // cache refresh timeout in seconds
+    private long refreshTimeout = 15; // cache refresh timeout in seconds
 
     private PreemptiveRefresh preemptiveRefresh = new PreemptiveRefresh();
 
@@ -20,11 +20,11 @@ public class JwtClientCache {
         this.enabled = enabled;
     }
 
-    public int getMinimumTimeToLive() {
+    public long getMinimumTimeToLive() {
         return minimumTimeToLive;
     }
 
-    public void setMinimumTimeToLive(int minimumTimeToLive) {
+    public void setMinimumTimeToLive(long minimumTimeToLive) {
         this.minimumTimeToLive = minimumTimeToLive;
     }
 
@@ -36,11 +36,11 @@ public class JwtClientCache {
         this.preemptiveRefresh = preemptiveRefresh;
     }
 
-    public int getRefreshTimeout() {
+    public long getRefreshTimeout() {
         return refreshTimeout;
     }
 
-    public void setRefreshTimeout(int refreshTimeout) {
+    public void setRefreshTimeout(long refreshTimeout) {
         this.refreshTimeout = refreshTimeout;
     }
 }
